feat: limit the number of subtree nodes
