+++ app/mock/policyholder-system.json
@@ -17,7 +17,7 @@
       "responses": [
         {
           "uuid": "32f22824-79eb-46aa-9da6-4fb2a897b283",
-          "body": "{\n  \"code\": \"{{queryParam 'code' '43bda7c2-bb84-4ace-9704-efcef436065b'}}\",\n  \"name\": \"{{faker 'person.firstName'}} {{faker 'person.lastName'}}\",\n  \"registration_date\": \"{{date '2023-11-20' '2024-08-25'}}\",\n  \"introducer_code\": \"{{faker 'string.uuid'}}\",\n  \"l\": [\n    {{# repeat 5 }}\n      {\n        \"code\": \"{{faker 'string.uuid'}}\",\n        \"name\": \"{{faker 'person.firstName'}} {{faker 'person.lastName'}}\",\n        \"registration_date\": \"{{date '2023-11-20' '2024-08-25'}}\",\n        \"introducer_code\": \"{{oneOf (array queryParam 'code' '38622bda-5f9a-4c38-8d95-3747cfd96453' '43bda7c2-bb84-4ace-9704-efcef436065b')}}\"\n      },\n    {{/ repeat }},\n    { \n      \"code\": \"{{faker 'string.uuid'}}\",\n      \"name\": \"{{faker 'person.firstName'}} {{faker 'person.lastName'}}\",\n      \"registration_date\": \"{{date '2023-11-20' '2024-08-25'}}\",\n      \"introducer_code\": \"{{queryParam 'code'}}\"\n    }\n  ],\n  \"r\": [\n    {{# repeat 6 }}\n      {\n        \"code\": \"{{faker 'string.uuid'}}\",\n        \"name\": \"{{faker 'person.firstName'}} {{faker 'person.lastName'}}\",\n        \"registration_date\": \"{{date '2023-11-20' '2024-08-25'}}\",\n        \"introducer_code\": \"{{oneOf (array 'a625a2a4-b586-47a5-a8fa-ced4ba763895' 'fcfa1851-9752-49b6-b954-e6fd8c26de18' '43bda7c2-bb84-4ace-9704-efcef436065b')}}\"\n      }\n    {{/ repeat }}\n  ]\n}",
+          "body": "{\n  \"code\": \"{{queryParam 'code' '43bda7c2-bb84-4ace-9704-efcef436065b'}}\",\n  \"name\": \"{{faker 'person.firstName'}} {{faker 'person.lastName'}}\",\n  \"registration_date\": \"{{date '2023-11-20' '2024-08-25'}}\",\n  \"introducer_code\": \"{{faker 'string.uuid'}}\",\n  \"l\": [\n    {{# repeat 15 }}\n      {\n        \"code\": \"{{faker 'string.uuid'}}\",\n        \"name\": \"{{faker 'person.firstName'}} {{faker 'person.lastName'}}\",\n        \"registration_date\": \"{{date '2023-11-20' '2024-08-25'}}\",\n        \"introducer_code\": \"{{oneOf (array queryParam 'code' '38622bda-5f9a-4c38-8d95-3747cfd96453' '43bda7c2-bb84-4ace-9704-efcef436065b')}}\"\n      },\n    {{/ repeat }},\n    { \n      \"code\": \"{{faker 'string.uuid'}}\",\n      \"name\": \"{{faker 'person.firstName'}} {{faker 'person.lastName'}}\",\n      \"registration_date\": \"{{date '2023-11-20' '2024-08-25'}}\",\n      \"introducer_code\": \"{{queryParam 'code'}}\"\n    }\n  ],\n  \"r\": [\n    {{# repeat 6 }}\n      {\n        \"code\": \"{{faker 'string.uuid'}}\",\n        \"name\": \"{{faker 'person.firstName'}} {{faker 'person.lastName'}}\",\n        \"registration_date\": \"{{date '2023-11-20' '2024-08-25'}}\",\n        \"introducer_code\": \"{{oneOf (array 'a625a2a4-b586-47a5-a8fa-ced4ba763895' 'fcfa1851-9752-49b6-b954-e6fd8c26de18' '43bda7c2-bb84-4ace-9704-efcef436065b')}}\"\n      }\n    {{/ repeat }}\n  ]\n}",
           "latency": 0,
           "statusCode": 200,
           "label": "",
